# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QA Playground REST API",
    "version": "1.0.0",
    "description": "REST API designed for practicing manual, API, automation, and performance testing. Includes comprehensive error handling, JWT auth with refresh tokens, role-based access control, cart, products, and order management.",
    "contact": {
      "name": "QA Playground Team",
      "email": "support@qaplayground.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/auth/login or /api/auth/register"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "example": "Premium noise-cancelling over-ear headphones"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 79.99
          },
          "stock": {
            "type": "integer",
            "example": 150
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/p-001.jpg"
          },
          "image_url": {
            "type": "string",
            "example": "/images/products/p-001.jpg"
          },
          "created_at": {
            "type": "string",
            "example": "2026-10-02 04:08:24"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-10-02 04:08:24"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "List products (Public; pagination, search, sort, filter)",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 12
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 12
                        },
                        "total": {
                          "type": "integer",
                          "example": 50
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price",
                  "category"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Wireless Ergonomic Keyboard"
                  },
                  "description": {
                    "type": "string",
                    "example": "Split mechanical ergonomic keyboard."
                  },
                  "price": {
                    "type": "number",
                    "example": 149.99
                  },
                  "stock": {
                    "type": "integer",
                    "example": 45
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "Electronics",
                      "Clothing",
                      "Books",
                      "Home & Kitchen",
                      "Sports"
                    ],
                    "example": "Electronics"
                  },
                  "imageUrl": {
                    "type": "string",
                    "description": "Relative /images/... path or http(s) URL (optional, defaults to /images/placeholder.svg)",
                    "example": "/images/products/p-001.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Product created successfully"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Admin access required"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get product by ID (Public)",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "summary": "Update product by ID (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "category": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "description": "Relative /images/... path or http(s) URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Product updated successfully"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Product not found"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "summary": "Delete product by ID (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully (no content)"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  }
}